+++ jeap-archrepo-persistence/src/main/resources/db/migration/common/V2_0_1__fix_variants.sql
@@ -1,0 +1,3 @@
+UPDATE message_graph
+SET variant = SUBSTRING(variant FROM POSITION('/' IN variant) + 1)
+WHERE POSITION('/' IN variant) > 0;