+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.0] - 2025-11-21
+
+### Changed
+
+- Fixed variant extraction logic. 
+- Optimized exception handling for 404 status codes, when graphs are not found for a component. 
+
 ## [4.1.0] - 2025-11-19
 
 ### Changed